# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pmagent API",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/auth/device/code": {
      "post": {
        "tags": [
          "device login"
        ],
        "summary": "Start Device Login",
        "description": "Step 1 (CLI): get a `device_code` to keep and a `user_code` to show the user, with the\npage where they approve it. Valid 10 minutes.",
        "operationId": "start_device_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request body or parameters failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/device/token": {
      "post": {
        "tags": [
          "device login"
        ],
        "summary": "Exchange Device Code",
        "description": "Step 3 (CLI): poll every `interval` seconds. Until the user decides, returns 400 with\n`type` ending in `authorization_pending` (keep polling), `slow_down` (poll less often),\n`access_denied` or `expired_token` (stop). On approval, returns an API token once.",
        "operationId": "exchange_device_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTokenCreated"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, for example an invalid or expired link or code",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request body or parameters failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/device/lookup": {
      "post": {
        "tags": [
          "device login"
        ],
        "summary": "Lookup Device",
        "description": "Step 2 (web app): show what's asking for access before the user approves.",
        "operationId": "lookup_device",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceLookup"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role or token scope doesn't allow this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to you",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request body or parameters failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/auth/device/approve": {
      "post": {
        "tags": [
          "device login"
        ],
        "summary": "Approve Device",
        "description": "Step 2 (web app): approve. Needs a login session, not an API token.",
        "operationId": "approve_device",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role or token scope doesn't allow this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to you",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request body or parameters failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/auth/device/deny": {
      "post": {
        "tags": [
          "device login"
        ],
        "summary": "Deny Device",
        "description": "Step 2 (web app): deny. The CLI's next poll gets `access_denied`.",
        "operationId": "deny_device",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role or token scope doesn't allow this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to you",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request body or parameters failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/me/tokens": {
      "get": {
        "tags": [
          "api tokens"
        ],
        "summary": "List Tokens",
        "description": "Your active API tokens. The token values themselves are never shown again.",
        "operationId": "list_tokens",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApiTokenRead"
                  },
                  "type": "array",
                  "title": "Response List Tokens"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "api tokens"
        ],
        "summary": "Create Token",
        "description": "Create an API token. It's in this response only; store it now. Needs a login\nsession: an API token can't create another (403).",
        "operationId": "create_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiTokenCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTokenCreated"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role or token scope doesn't allow this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Request body or parameters failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/me/tokens/{token_id}": {
      "delete": {
        "tags": [
          "api tokens"
        ],
        "summary": "Revoke Token",
        "description": "Any of your credentials can revoke your tokens, including the token itself (CLI logout).",
        "operationId": "revoke_token",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid, or expired credentials",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to you",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiTokenCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "scopes": {
            "items": {
              "$ref": "#/components/schemas/Scope"
            },
            "type": "array",
            "minItems": 1,
            "title": "Scopes",
            "default": [
              "read",
              "write"
            ]
          },
          "expires_in_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 365.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In Days",
            "default": 90
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ApiTokenCreate"
      },
      "ApiTokenCreated": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "display_prefix": {
            "type": "string",
            "title": "Display Prefix"
          },
          "scopes": {
            "items": {
              "$ref": "#/components/schemas/Scope"
            },
            "type": "array",
            "title": "Scopes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "token_type": {
            "type": "string",
            "const": "bearer",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "display_prefix",
          "scopes",
          "created_at",
          "expires_at",
          "last_used_at",
          "token"
        ],
        "title": "ApiTokenCreated",
        "description": "Returned once, at creation. The token can't be retrieved again."
      },
      "ApiTokenRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "display_prefix": {
            "type": "string",
            "title": "Display Prefix"
          },
          "scopes": {
            "items": {
              "$ref": "#/components/schemas/Scope"
            },
            "type": "array",
            "title": "Scopes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "display_prefix",
          "scopes",
          "created_at",
          "expires_at",
          "last_used_at"
        ],
        "title": "ApiTokenRead"
      },
      "DeviceCodeRequest": {
        "properties": {
          "client_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Client Name",
            "default": "pmagent CLI"
          },
          "scopes": {
            "items": {
              "$ref": "#/components/schemas/Scope"
            },
            "type": "array",
            "minItems": 1,
            "title": "Scopes",
            "default": [
              "read",
              "write"
            ]
          }
        },
        "type": "object",
        "title": "DeviceCodeRequest"
      },
      "DeviceCodeResponse": {
        "properties": {
          "device_code": {
            "type": "string",
            "title": "Device Code"
          },
          "user_code": {
            "type": "string",
            "title": "User Code"
          },
          "verification_uri": {
            "type": "string",
            "title": "Verification Uri"
          },
          "verification_uri_complete": {
            "type": "string",
            "title": "Verification Uri Complete"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "interval": {
            "type": "integer",
            "title": "Interval"
          }
        },
        "type": "object",
        "required": [
          "device_code",
          "user_code",
          "verification_uri",
          "verification_uri_complete",
          "expires_in",
          "interval"
        ],
        "title": "DeviceCodeResponse"
      },
      "DeviceLookup": {
        "properties": {
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "scopes": {
            "items": {
              "$ref": "#/components/schemas/Scope"
            },
            "type": "array",
            "title": "Scopes"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "client_name",
          "scopes",
          "expires_at"
        ],
        "title": "DeviceLookup"
      },
      "DeviceTokenRequest": {
        "properties": {
          "device_code": {
            "type": "string",
            "maxLength": 256,
            "title": "Device Code"
          }
        },
        "type": "object",
        "required": [
          "device_code"
        ],
        "title": "DeviceTokenRequest"
      },
...
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "examples": [
              "https://pmagent.dev/problems/not_found"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "examples": [
              "Not Found"
            ]
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "examples": [
              404
            ]
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Workspace not found"
            ]
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "examples": [
              "3f2b9c0e8a4d4e8f9b1c2d3e4f5a6b7c"
            ]
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Field errors; only on 422 validation errors."
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetail",
        "description": "RFC 9457 problem details. Every error response has this shape."
      },
...
      "Scope": {
        "type": "string",
        "enum": [
          "read",
          "write"
        ],
        "title": "Scope"
      },
...
      "UserCodeRequest": {
        "properties": {
          "user_code": {
            "type": "string",
            "maxLength": 16,
            "title": "User Code"
          }
        },
        "type": "object",
        "required": [
          "user_code"
        ],
        "title": "UserCodeRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token from /v1/auth/login, or an API token (pmat_...)",
        "scheme": "bearer"
      }
    }
  },
...
}